+++ lua/github-theme/types/init.lua
@@ -1,0 +1,5 @@
+local types = {}
+
+types.gt = require("github-theme.types.gt")
+
+return types